feat(heaader): add header commands
This brings in the `--header` and `--header-last` commands to put a
header at either the top of the output or the end. This lets you then
pass that header onward to `fzf` to make your output just a little bit
prettier.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>