feat(hostname): support hostname based filtering
This brings in support for fancy hostname based filtering, which is the
last feature that I needed to replicate atuin support -- or at least get
it close enough. This is probably enough for a 0.1.0 release once docs
are there, but not perfect because atuin caches commands before writing
them to sqlite.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>